Fixed alliance trajectories, Updated model, Updated detection logic based on L or R starting position (untested)
