+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/shared/GlobalState.java
@@ -1,0 +1,9 @@
+package org.firstinspires.ftc.teamcode.shared;
+
+public class GlobalState {
+
+    public static enum ALLIANCE_POS {
+        LEFT,
+        RIGHT
+    }
+}
